DATAMONGO-2393 - Polishing.
Move query rewriting into CountQuery. Consider existing $and items during query rewrite.
